pyproject.toml
[project]
name = "vk-oc-projet05-deploiml"
version = "0.1.0"
description = "Configuration de l'environnement virtuel pour réaliser l'ensemble du projet"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Data / ML
    "imbalanced-learn>=0.14.1",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "missingno>=0.5.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pyarrow>=23.0.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "statsmodels>=0.14.6",
    "lightgbm>=4.0.0",
    # DB
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    # Tests
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "bcrypt>=5.0.0",
    "pyjwt>=2.12.1",
    "pydantic-settings>=2.13.1",
]
